Added Player2 controller, started refactoring
I created a second controller for Player2, using the inputs I, J, L, the apostrophe/quote key, and U, but now I need to refactor the code so that the Player1 controller, the Player2 controller, and the CPU use the same base script that can dynamically change its target/controls based on who's who. Just saving the headache of having to change 3 300+ line size scripts on every edit.